+++ src/main/scala/sample/hellopond/StringFish.scala
@@ -3,7 +3,7 @@ package sample.hellopond
 import scala.collection.immutable.{Seq => ISeq}
 
 // Fish that reacts to StringEvents, does not handle commands
-case class StringEvent(override val timestamp: Int, override val source: String, value: String)
+case class StringEvent(override val timestamp: Long, override val source: String, value: String)
     extends Event(timestamp, source)
 
 object StringFish {
